+++ R/nettable.R
@@ -37,6 +37,8 @@
 #'   should be back transformed. For example, if \code{backtransf =
 #'   TRUE}, results for \code{sm = "OR"} are printed as odds ratios
 #'   rather than log odds ratios.
+#' @param nchar.trts A numeric defining the minimum number of
+#'   characters used to create unique treatment names.
 #' @param digits Minimal number of significant digits, see
 #'   \code{print.default}.
 #' @param digits.I2 Minimal number of significant digits for I-squared
@@ -72,6 +74,8 @@
 #'   file.
 #' @param overwrite A logical indicating whether an existing Excel
 #'   file should be overwritten.
+#' @param legend A logical indicating whether a legend should be
+#'   printed for abbreviated treatment names.
 #' @param warn A logical indicating whether warnings should be
 #'   printed.
 #' @param verbose A logical indicating whether progress information
@@ -201,6 +205,8 @@ nettable <- function(...,
                      ##
                      backtransf = NULL,
                      ##
+                     nchar.trts = if (writexl) 666 else NULL,
+                     ##
                      digits = gs("digits"),
                      digits.I2 = gs("digits.I2"),
                      digits.pval = gs("digits.pval"),
@@ -324,6 +330,18 @@ nettable <- function(...,
   ##
   if (length(unique(backtransfs)) == 1)
     backtransfs <- unique(backtransfs)
+  ##
+  if (is.null(nchar.trts)) {
+    nchar.trts <- vector("numeric", n.netmeta)
+    for (i in n.i)
+      nchar.trts[i] <- args[[i]]$nchar.trts
+    ##
+    if (length(unique(nchar.trts)) == 1)
+      nchar.trts <- unique(nchar.trts)
+    else
+      nchar.trts <- min(nchar.trts, na.rm = TRUE)
+  }
+  chknumeric(nchar.trts, min = 1, length = 1)
   
   
   ##
@@ -425,7 +443,29 @@ nettable <- function(...,
   else
     table.random$n <- formatN(round(table.random$n), digits = 0,
                               text.NA = text.NA, big.mark = big.mark)
-  
+  ##
+  trts <- c(table.common[["Arm 1"]], table.common[["Arm 2"]])
+  ##
+  if (!(all(nchar.trts > nchar(trts)))) {
+    trts.abbr <- treats(trts, nchar.trts)
+    ##
+    table.common[["Arm 1"]] <-
+      as.character(factor(table.common[["Arm 1"]],
+                          levels = trts, labels = trts.abbr))
+    table.common[["Arm 2"]] <-
+      as.character(factor(table.common[["Arm 2"]],
+                          levels = trts, labels = trts.abbr))
+    ##
+    table.random[["Arm 1"]] <-
+      as.character(factor(table.random[["Arm 1"]],
+                          levels = trts, labels = trts.abbr))
+    table.random[["Arm 2"]] <-
+      as.character(factor(table.random[["Arm 2"]],
+                          levels = trts, labels = trts.abbr))
+  }
+  else
+    trts.abbr <- trts
+
   
   ##
   ##
@@ -504,6 +544,10 @@ nettable <- function(...,
               sm = sms,
               level.ma = unique(levs),
               ##
+              nchar.trts,
+              trts = unique(trts),
+              trts.abbr = unique(trts.abbr),
+              ##
               version = packageDescription("netmeta")$Version
               )
   ##
@@ -526,7 +570,8 @@ nettable <- function(...,
 #' @export
 
 
-print.nettable <- function(x, common = x$x$common, random = x$x$random, ...) {
+print.nettable <- function(x, common = x$x$common, random = x$x$random,
+                           legend = TRUE, ...) {
   
   ##
   ##
@@ -545,6 +590,7 @@ print.nettable <- function(x, common = x$x$common, random = x$x$random, ...) {
   common <- deprecated(common, missing(common), args, "fixed", FALSE)
   chklogical(common)
   chklogical(random)
+  chklogical(legend)
   
   
   ##
@@ -624,6 +670,14 @@ print.nettable <- function(x, common = x$x$common, random = x$x$random, ...) {
     }
   }
   
+  
+  ##
+  ##
+  ## (4) Legend
+  ##
+  ##
+  if (!is.null(x$trts) && !is.null(x$trts.abbr))
+    legendabbr(x$trts, x$trts.abbr, legend)
   
   invisible(NULL)
 }